pyproject.toml
[tool.poetry]
name = "cryptopals"
version = "0.1.0"
description = ""
authors = ["Aidan Williams <aweraw@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pycryptodome = "^3.10.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
